requirements.txt
numpy==2.1.2
opencv_contrib_python==4.10.0.84
opencv_python==4.10.0.84
opencv_python_headless==4.10.0.84
Pillow==11.0.0
torch==2.2.1
